Less Jitter for ALSA Playback Plugin
Originally the ALSA playback plugin will cause a lot of jitter for
pipe flushes because the application is not guarenteed to flush
a block every period and multiple blocks can be flushed at the same
time instead.

The problem can cause drain operation to buffer underrun when the
latency is lower.

The commit change the playback plugin and uses a separate thread
to flush blocks every period.